+++ app/src/main/java/goalkeeper/matheus/goalkeeper/jogadas/ReporMaoTela.java
@@ -8,6 +8,7 @@ import android.view.View;
 import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.CheckBox;
+import android.widget.EditText;
 import android.widget.Spinner;
 
 import java.util.ArrayList;
@@ -41,20 +42,23 @@ public class ReporMaoTela extends JogadaOfensivaTela {
         mSpinPrimeiraBola = (Spinner) findViewById(R.id.spinner_primBola);
         mSpinSegundaBola = (Spinner) findViewById(R.id.spinner_segBola);
         mCheckErrou = (CheckBox) findViewById(R.id.check_erroJO);
-        mSpinErro = (Spinner) findViewById(R.id.spinner_erroJO);
+       // mSpinErro = (Spinner) findViewById(R.id.spinner_erroJO);
+        mTextErro = (EditText) findViewById(R.id.edit_txt_observacao);
         btnSalvarJO = (Button) findViewById(R.id.btn_salvarJO);
         btnCancelarJO = (Button) findViewById(R.id.btn_calcelJO);
 
+        /*
         mSpinErro = (Spinner) findViewById(R.id.spinner_erroJO);
         mCheckErrou = (CheckBox) findViewById(R.id.check_erroJO);
-        mSpinErro.setVisibility(View.GONE);
+        mTextErro.setVisibility(View.GONE);
         mCheckErrou.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                if(mCheckErrou.isChecked()) mSpinErro.setVisibility(View.VISIBLE);
-                else mSpinErro.setVisibility(View.GONE);
+                if(mCheckErrou.isChecked()) mTextErro.setVisibility(View.VISIBLE);
+                else mTextErro.setVisibility(View.GONE);
             }
         });
+        */
 
         carregarValores();
 
@@ -66,7 +70,7 @@ public class ReporMaoTela extends JogadaOfensivaTela {
                     int idJogadaOfensiva=saveJO();
                     mDb.cadastrarReporMao(idJogadaOfensiva, mSpinTipoReporMao.getSelectedItem().toString());
                     if(errou==1)CadastroPartida.historico += "REPOSIÇÃO COM MÃO:\n"+
-                            tempo+" minutos, reposição foi do tipo "+ mSpinTipoReporMao.getSelectedItem().toString()+ ", bola foi no setor "+setorBolaFoi+ ", primeira bola ganha por "+primeiraBola+ ", segunda bola ganha por "+segundaBola+ "\nErro: "+erro+"\n\n";
+                            tempo+" minutos, reposição foi do tipo "+ mSpinTipoReporMao.getSelectedItem().toString()+ ", bola foi no setor "+setorBolaFoi+ ", primeira bola ganha por "+primeiraBola+ ", segunda bola ganha por "+segundaBola+ "\nErro: "+observacao+"\n\n";
                     if(errou==0)CadastroPartida.historico += "REPOSIÇÃO COM MÃO:\n"+
                             tempo+" minutos, reposição foi do tipo "+ mSpinTipoReporMao.getSelectedItem().toString()+ ", bola foi no setor "+setorBolaFoi+ ", primeira bola ganha por "+primeiraBola+ ", segunda bola ganha por "+segundaBola+ "\nAcertou a jogada\n\n";
 
@@ -83,7 +87,6 @@ public class ReporMaoTela extends JogadaOfensivaTela {
                 finish();
             }
         });
-
     }
 
     private void carregarValores() {
@@ -137,6 +140,7 @@ public class ReporMaoTela extends JogadaOfensivaTela {
         ArrayAdapter<String> adapter4 = new ArrayAdapter<String>(this, android.R.layout.simple_spinner_dropdown_item, arraySegBola);
         mSpinSegundaBola.setAdapter(adapter4);
 
+        /*
         arrayErros = new ArrayList<String>();
         arrayErros.add("Selecione o erro");
         arrayErros.add("muito baixo");
@@ -145,8 +149,8 @@ public class ReporMaoTela extends JogadaOfensivaTela {
         arrayErros.add("erro de direção");
         ArrayAdapter<String> adapter5 = new ArrayAdapter<String>(this, android.R.layout.simple_spinner_dropdown_item, arrayErros);
         mSpinErro.setAdapter(adapter5);
+        */
     }
-
 
     public void mensagem() {
         AlertDialog alertDialog = new AlertDialog.Builder(this).create();
@@ -158,7 +162,4 @@ public class ReporMaoTela extends JogadaOfensivaTela {
         });
         alertDialog.show();
     }
-
-
-
 }
