+++ app/src/main/java/model/CaraCara.java
@@ -7,7 +7,7 @@ public class CaraCara extends JogadaDefensiva{
     private int idJogadaDefensiva;
     private int tipoCaraCara;
 
-    public CaraCara(int idPartida, int tempo, String setorBolaVeio, String setorBolaFoi, int errou, String erro, String tipoFinalizacao, int gol) {
-        super(idPartida, tempo, setorBolaVeio, setorBolaFoi, errou, erro, tipoFinalizacao, gol);
+    public CaraCara(int idPartida, int tempo, String setorBolaVeio, String setorBolaFoi, int errou, String observacao, String tipoFinalizacao, int gol) {
+        super(idPartida, tempo, setorBolaVeio, setorBolaFoi, errou, observacao, tipoFinalizacao, gol);
     }
 }
