+++ app/src/main/java/model/Cruzamento.java
@@ -11,7 +11,7 @@ public class Cruzamento extends JogadaDefensiva{
     private String erroCruzamento;
     private int idJogadaDefensiva;
 
-    public Cruzamento(int idPartida, int tempo, String setorBolaVeio, String setorBolaFoi, int errou, String erro, String tipoFinalizacao, int gol) {
-        super(idPartida, tempo, setorBolaVeio, setorBolaFoi, errou, erro, tipoFinalizacao, gol);
+    public Cruzamento(int idPartida, int tempo, String setorBolaVeio, String setorBolaFoi, int errou, String observacao, String tipoFinalizacao, int gol) {
+        super(idPartida, tempo, setorBolaVeio, setorBolaFoi, errou, observacao, tipoFinalizacao, gol);
     }
 }
