+++ app/src/main/java/model/DefBase.java
@@ -6,7 +6,7 @@ package model;
 public class DefBase extends JogadaDefensiva{
     private int idJogadaDefensiva;
 
-    public DefBase(int idPartida, int tempo, String setorBolaVeio, String setorBolaFoi, int errou, String erro, String tipoFinalizacao, int gol) {
-        super(idPartida, tempo, setorBolaVeio, setorBolaFoi, errou, erro, tipoFinalizacao, gol);
+    public DefBase(int idPartida, int tempo, String setorBolaVeio, String setorBolaFoi, int errou, String observacao, String tipoFinalizacao, int gol) {
+        super(idPartida, tempo, setorBolaVeio, setorBolaFoi, errou, observacao, tipoFinalizacao, gol);
     }
 }
