+++ app/src/main/java/model/DefPe.java
@@ -6,7 +6,7 @@ package model;
 public class DefPe extends JogadaDefensiva{
     private int idJogadaDefensiva;
 
-    public DefPe(int idPartida, int tempo, String setorBolaVeio, String setorBolaFoi, int errou, String erro, String tipoFinalizacao, int gol) {
-        super(idPartida, tempo, setorBolaVeio, setorBolaFoi, errou, erro, tipoFinalizacao, gol);
+    public DefPe(int idPartida, int tempo, String setorBolaVeio, String setorBolaFoi, int errou, String observacao, String tipoFinalizacao, int gol) {
+        super(idPartida, tempo, setorBolaVeio, setorBolaFoi, errou, observacao, tipoFinalizacao, gol);
     }
 }
