+++ app/src/main/java/model/DefPunho.java
@@ -7,7 +7,7 @@ public class DefPunho extends JogadaDefensiva{
     private int idJogadaDefensiva;
     private String tipoDefesaPunho;
 
-    public DefPunho(int idPartida, int tempo, String setorBolaVeio, String setorBolaFoi, int errou, String erro, String tipoFinalizacao, int gol) {
-        super(idPartida, tempo, setorBolaVeio, setorBolaFoi, errou, erro, tipoFinalizacao, gol);
+    public DefPunho(int idPartida, int tempo, String setorBolaVeio, String setorBolaFoi, int errou, String observacao, String tipoFinalizacao, int gol) {
+        super(idPartida, tempo, setorBolaVeio, setorBolaFoi, errou, observacao, tipoFinalizacao, gol);
     }
 }
