+++ app/src/main/java/model/TiroMeta.java
@@ -7,7 +7,7 @@ public class TiroMeta extends JogadaOfensiva{
     private String tipoTiroMeta;
     private int idJogadaOfensiva;
 
-    public TiroMeta(int idPartida, int tempo, String setorBolaFoi, String primeiraBola, String segundaBola, int errou, String erro) {
-        super(idPartida, tempo, setorBolaFoi, primeiraBola, segundaBola, errou, erro);
+    public TiroMeta(int idPartida, int tempo, String setorBolaFoi, String primeiraBola, String segundaBola, int errou, String observacao) {
+        super(idPartida, tempo, setorBolaFoi, primeiraBola, segundaBola, errou, observacao);
     }
 }
