+++ app/src/main/res/layout/activity_jogada_defensiva_tela.xml
@@ -33,10 +33,14 @@
         android:id="@+id/check_erro"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content" />
-    <Spinner
-        android:id="@+id/spinner_erroJD"
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Observação:"/>
+    <EditText
+        android:id="@+id/edit_txt_observacao"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"></Spinner>
+        android:layout_height="wrap_content" />
     <LinearLayout
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
