+++ app/src/main/res/layout/activity_jogada_ofensiva_tela.xml
@@ -28,12 +28,16 @@
         android:text="Houve Erro do Goleiro?"
         android:id="@+id/check_erroJO"
         android:layout_width="wrap_content"
-        android:layout_height="wrap_content" />
-    <Spinner
-        android:id="@+id/spinner_erroJO"
-        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:paddingBottom="@dimen/activity_vertical_margin"></Spinner>
+        android:checked="false" />
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Observação:"/>
+    <EditText
+        android:id="@+id/edit_txt_observacao"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content" />
     <LinearLayout
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
