+++ local.properties
@@ -7,5 +7,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Mon Oct 31 11:27:39 BRST 2016
-sdk.dir=C\:\\Users\\Matheus\\AppData\\Local\\Android\\Sdk
+#Wed Nov 02 18:37:06 BRST 2016
+sdk.dir=/home/kenneth/Android/Sdk
